CRT-396 address review comments
