+++ test/e2e/usersignup_test.go
@@ -2,8 +2,6 @@ package e2e
 
 import (
 	"context"
-	"crypto/md5"
-	"encoding/hex"
 	"testing"
 	"time"
 
@@ -170,10 +168,6 @@ func newUserSignup(t *testing.T, host *wait.HostAwaitility, username string, ema
 	require.NoError(t, err)
 	require.True(t, ok)
 
-	md5hash := md5.New()
-	_, _ = md5hash.Write([]byte(email))
-	emailHash := hex.EncodeToString(md5hash.Sum(nil))
-
 	return &v1alpha1.UserSignup{
 		ObjectMeta: v1.ObjectMeta{
 			Name:      uuid.NewV4().String(),
@@ -182,7 +176,7 @@ func newUserSignup(t *testing.T, host *wait.HostAwaitility, username string, ema
 				v1alpha1.UserSignupUserEmailAnnotationKey: email,
 			},
 			Labels: map[string]string{
-				v1alpha1.UserSignupUserEmailHashLabelKey: emailHash,
+				v1alpha1.UserSignupUserEmailHashLabelKey: CalcMD5(email),
 			},
 		},
 		Spec: v1alpha1.UserSignupSpec{
