+++ wait/host.go
@@ -2,13 +2,12 @@ package wait
 
 import (
 	"context"
-	"crypto/md5"
-	"encoding/hex"
 	"reflect"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
 	"github.com/codeready-toolchain/toolchain-common/pkg/test"
+	"github.com/codeready-toolchain/toolchain-e2e/test/e2e"
 
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
@@ -131,11 +130,7 @@ func (a *HostAwaitility) WaitForUserSignup(name string, criteria ...UserSignupWa
 
 // WaitForBannedUser waits until there is a BannedUser available with the given email
 func (a *HostAwaitility) WaitForBannedUser(email string) (bannedUser *toolchainv1alpha1.BannedUser, err error) {
-	md5hash := md5.New()
-	// Ignore the error, as this implementation cannot return one
-	_, _ = md5hash.Write([]byte(email))
-	emailHash := hex.EncodeToString(md5hash.Sum(nil))
-	labels := map[string]string{toolchainv1alpha1.BannedUserEmailHashLabelKey: emailHash}
+	labels := map[string]string{toolchainv1alpha1.BannedUserEmailHashLabelKey: e2e.CalcMd5(email)}
 	opts := client.MatchingLabels(labels)
 
 	err = wait.Poll(a.RetryInterval, a.Timeout, func() (done bool, err error) {
